clean up default settings
